factor out OrderListener classes
