+++ univocity-trader-marketsimulator/pom.xml
@@ -35,6 +35,11 @@
 			<artifactId>univocity-trader-binance</artifactId>
 			<version>${revision}</version>
 		</dependency>
+ 		<dependency>
+			<groupId>com.univocity</groupId>
+			<artifactId>univocity-trader-simplestrategystatistics</artifactId>
+			<version>${revision}</version>
+		</dependency>		
 		<dependency>
 			<groupId>commons-cli</groupId>
 			<artifactId>commons-cli</artifactId>
