+++ univocity-trader-orderlisteners/pom.xml
@@ -1,0 +1,26 @@
+<!--
+  ~ Copyright (c) 2013 Univocity Software Pty Ltd. All rights reserved.
+  ~ This file is subject to the terms and conditions defined in file
+  ~ 'LICENSE.txt', which is part of this source code package.
+  -->
+
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+
+	<modelVersion>4.0.0</modelVersion>
+
+	<parent>
+		<groupId>com.univocity</groupId>
+		<artifactId>univocity-trader</artifactId>
+		<version>${revision}</version>
+	</parent>
+
+	<artifactId>univocity-trader-orderlisteners</artifactId>
+
+	<name>univocity-trader-orderlisteners</name>
+	<packaging>pom</packaging>
+
+	<modules>
+		<module>univocity-trader-simplestrategystatistics</module>
+	</modules>
+</project>
