+++ univocity-trader-orderlisteners/univocity-trader-simplestrategystatistics/pom.xml
@@ -1,0 +1,25 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+
+	<modelVersion>4.0.0</modelVersion>
+
+	<parent>
+		<groupId>com.univocity</groupId>
+		<artifactId>univocity-trader-orderlisteners</artifactId>
+		<version>${revision}</version>
+	</parent>
+
+	<artifactId>univocity-trader-simplestrategystatistics</artifactId>
+
+	<name>univocity-trader-simplestrategystatistics</name>
+	<packaging>jar</packaging>
+
+	<dependencies>
+		<dependency>
+			<groupId>com.univocity</groupId>
+			<artifactId>univocity-trader-core</artifactId>
+			<version>${revision}</version>
+		</dependency>
+	</dependencies>
+
+</project>
